feat: add persona system and skill packages
- Add personas for Confucius, Laozi, and Zen
- Add personaContent utility
- Add speak-chan and speak-dadao skill packages
- Add laozi and zen assembly images
- Update translate API, layout, page styles
- Update cardDownload and prompt utilities
- Update configuration files
