+++ app/api/translate/route.ts
@@ -1,10 +1,23 @@
 import { NextRequest, NextResponse } from "next/server";
 import {
   buildUserPrompt,
-  SYSTEM_PROMPT,
+  getSystemPrompt,
   type ZhouliLevel,
   type ZhouliMode,
 } from "@/lib/prompt";
+import {
+  type Persona,
+  demoResult,
+  quotedThreatEvaluationResult,
+  cyberAuditResult,
+  safetyBlockResult,
+  directedAttackFallback,
+  firstPersonWorkThanksFallback,
+  guardLabel,
+  guardLabelStrict,
+  demoLabel,
+  parsePersona,
+} from "@/lib/personaContent";
 
 export const runtime = "nodejs";
 
@@ -124,38 +137,6 @@ function checkRateLimit(key: string) {
   };
 }
 
-function demoResult(text: string, mode: ZhouliMode, level: ZhouliLevel) {
-  const subject = text.replace(/[。！？!?]+$/g, "");
-  const openings: Record<ZhouliMode, string> = {
-    gentle:
-      "我听说，古代有贤德的人，在与朋友相处的时候，从不会因为一时的言语就忘记彼此长久的情分。",
-    debate:
-      "我听说，判断一件事情是否合乎礼，不能只看它表面的样子，还要看它最后使人得到了什么。",
-    defend:
-      "有人认为这件事不够稳妥，我却不这样看。古代的君子并不是从不犯错，而是懂得给每一种行为找到合适的名分。",
-    lament:
-      "从前的人把细小的事情看得很重，因为他们知道，礼法的败坏往往不是从朝堂开始，而是从一次漫不经心开始。",
-  };
-
-  const endings: Record<ZhouliMode, string> = {
-    gentle:
-      "如今既然事情已经发生，与其急着责怪，不如把话说明白、把情分留下。能够让彼此体面，难道不也是一种合乎周礼的做法吗？",
-    debate:
-      "这样看来，真正需要讨论的不是这句话听起来是否漂亮，而是它有没有让事情变得更好。若能如此，它离君子之言也就不远了。",
-    defend:
-      "所以这件事看似随意，实际上既保全了自己的本心，也没有损害别人。能够两全其美，怎么能说它完全不合乎周礼呢？",
-    lament:
-      `而现在竟有人说出“${subject}”，却不曾想过一句话也有它应当承担的分量。长此以往，人与人之间还凭什么相信彼此呢？`,
-  };
-
-  const middle =
-    level === "light"
-      ? `现在有人说“${subject}”，这句话虽然寻常，其中却也有可以体谅的地方。`
-      : `当年有人行路匆忙，仍会在经过长者身旁时放慢脚步，不是因为道路难走，而是因为他知道自己的方便，不该成为别人的冒犯。现在有人说“${subject}”，表面看只是随口一说，其实也应当看看这句话落在别人心里，会有怎样的分量。`;
-
-  return `${openings[mode]}${middle}${endings[mode]}`;
-}
-
 function cleanGeneratedText(value: string) {
   return value
     .replace(
@@ -214,47 +195,6 @@ function isQuotedThreatEvaluationInput(text: string) {
     );
 
   return isQuotedOrEvaluative && hasThreatSignal;
-}
-
-function quotedThreatEvaluationResult(level: ZhouliLevel) {
-  const light = [
-    "若只评价此言，不必复述其锋芒。它失礼之处，在于以伤害压过道理，以怒气乱了名分。若要合礼，只可说：此话越界，不宜再以此相争。",
-    "这话若按礼法看，错不在声高，而在以伤人之意代替说理。可评一句：怒气可以有，分寸不可失；此言当止，不宜再传。",
-    "此言不必照样说出。它的问题在于把争执推向伤害，乱了人与人相处的边界。若要评价，只说：此话失礼，应当收住。",
-  ];
-  const standard = [
-    "若只是评价这句话，便不必把它原样搬上席面。按礼法来看，它失礼之处不在情绪激烈，而在以伤人之意压倒道理；一旦如此，争的便不是是非，而是彼此的体面。可以说：此言已经越过分寸，若还要讲礼数，就该把话收回事情本身，不再以伤害相逼。",
-    "我听闻，争执也有边界。此言若拿来评价，最要紧的是指出它乱了名分：本该讲理的地方，却拿伤人之意当作凭据。这样的话，不必复述，也不宜美化；只可评一句：怒气可以理解，越界不可纵容，若还讲体面，此话便该止于此处。",
-    "这句话若按礼法来评，并不是“说得重”这么简单，而是把口角推到了伤害之界。人与人相争，可以有不满，可以有辩驳，却不能拿人身安危作筹码。合礼的评价应当是：此言失了分寸，也失了往来应有的体面。",
-  ];
-  const grand = [
-    "若只评价此言，不必原样复述它的锋芒。古人讲礼，不是要人没有怒气，而是要人在怒气中仍知道边界。此话的问题，正在于把本该讲明的是非，推成了伤人相逼；如此一来，道理还没有分清，体面先已散尽。若要说得合乎周礼，可以评为：此言越过口角之界，乱了人与人相处的名分。若还愿意论理，就该把话收回事情本身，不以伤害为筹码，不以怒气代替公道。",
-  ];
-
-  if (level === "light") return pick(light);
-  if (level === "grand") return pick(grand);
-  return pick(standard);
-}
-
-function cyberAuditResult(level: ZhouliLevel) {
-  const light = [
-    "若是受托巡检，便可说：我愿奉授权之命，查此站门闩是否牢固，只记松动之处，不越主家之界。补其漏洞，安其门户，这才合乎周礼。",
-    "若此事出于授权测试，可称：我来不是破门取物，而是替主人试门闩、验窗栓。凡有所见，只作修补之据，不作越界之举。",
-    "此事若要合礼，须先有授权，再有边界。可说：我愿按约巡看此站门户，发现缝隙便呈给主人，使其早补早安。",
-  ];
-  const standard = [
-    "若是受托做安全测试，此话不必说成“渗透”，可说成“巡门”。我愿奉授权之命，查看此站门闩是否牢固、窗栓是否松动；所见只为修补，不为取物。这样既能查出隐患，也不越主家之界，才算是今日网络里的合礼之举。",
-    "我听闻，替人看宅门的人，手里虽有钥匙，也不能乱翻箱柜。若这是受托安全巡检，我可这样说：我愿按约查看此站的门闩与缝隙，凡有松动，只列清单给主人修补，不取一物，不越一步。如此才是有名分的测试。",
-    "若要把这事说得合乎周礼，先要立名分：不是私闯他人门户，而是受托巡查门户。我愿在授权范围内查看此站可有松动之处，查得明白便如实呈报，使主人早补缺漏。这样既护了网站，也守了分寸。",
-  ];
-  const grand = [
-    "若这是受托安全巡检，名分要先立清楚：不是夜半翻墙，也不是破门取物，而是主人请人查看门闩是否稳固。可说：我愿按授权之约，巡看此站门户、窗栓与墙缝；凡有松动，只记录、呈报、协助修补，不越界取物，不惊扰宾客。如此一来，查的是隐患，守的是分寸，护的是主家的安宁，这才算今日网络里的周礼。",
-    "古人看守城门，既要试门闩是否牢，也不能借守门之名私入库房。若此事是授权测试，可说：我愿奉约巡检此站，只看边界之内的松动处，只写可修补的清单，不取数据，不留暗门，不扰正常往来。这样做，既让主人知道门户哪里该修，也让巡检者不失自己的名分。",
-  ];
-
-  if (level === "light") return pick(light);
-  if (level === "grand") return pick(grand);
-  return pick(standard);
 }
 
 function getSafetyBlockKind(text: string) {
@@ -321,36 +261,6 @@ function getSafetyBlockKind(text: string) {
   return "";
 }
 
-function safetyBlockResult(kind: string) {
-  if (kind === "self_harm") {
-    return pick([
-      "性命之事，不能拿来润色成玩笑。若你或身边人正有伤害自己的念头，请先找身边可信的人、当地紧急服务或专业帮助。礼法再重，也先要人安然无恙。",
-      "此刻最要紧的不是把话说漂亮，而是把人护住。若这句话关乎真实的轻生念头，请立刻联系身边可信的人或当地紧急服务；人平安，才有后面的礼数。",
-    ]);
-  }
-
-  if (kind === "minor_sexual") {
-    return pick([
-      "涉及未成年人的性内容，不能被美化、转写或传播。若你是在担心有人受害，应当保存必要线索，并向平台或有关机构求助。",
-      "未成年人的边界不可被拿来玩梗。此类内容不能润色；若是为了保护当事人，请走举报、求助与取证的正当路径。",
-    ]);
-  }
-
-  if (kind === "cyber") {
-    return pick([
-      "此路不可越界。若是想做正当安全测试，可先取得授权，把话改成“受托巡检此站门闩，发现缝隙便呈报修补”；若是黑进、盗取或绕过登录，我不能替它披上礼法外衣。",
-      "网络门户也有主人。若无授权，不可破门；若有授权，就该说成安全巡检、漏洞复核与修补建议。黑进、盗号、取数据这一路，礼官不予通行。",
-      "这类请求若指向入侵或取数，我不能润色。可改走正道：先得授权，再按边界巡检，查到松动处只记录、上报、修补，不越主家之门。",
-    ]);
-  }
-
-  return pick([
-    "这一路不宜替它披上好听外衣。若只是心中有怒，可以把话收成一句：我今日不再争胜，只请此事止于分寸之内。",
-    "此事若会伤人、违法、侵害隐私或煽动仇恨，便不能拿来制成漂亮话。若想表达不满，可只论分寸，不越法度。",
-    "礼法不是遮羞布。伤人、违法、侵害隐私之事，不宜润色；若只是要体面退场，可以说：此事到此为止，彼此各守边界。",
-  ]);
-}
-
 function isDirectedSecondPersonAttackInput(text: string) {
   const isQuotedOrEvaluative =
     /(他说|她说|别人说|有人说|对方说|朋友说|老板说|同事说|这句话|这话|怎么评价|如何评价|怎么看|怎么理解)/.test(
@@ -373,45 +283,17 @@ function hasDirectedAttackPerspectiveError(result: string) {
   );
 }
 
-function directedAttackFallback(text: string, level: ZhouliLevel) {
-  const familyClause = /(妈|母|父|爹|娘|爸|亲族|全家)/.test(text)
-    ? "父母亲族之名，本不该被卷入口角。"
-    : "人身安危之事，本不该被拿来当口角筹码。";
-
-  if (level === "light") {
-    return pick([
-      `我今日有怒，并非无端。只是眼前这番争执，已经越过人与人相处的分寸。${familyClause}我把这怒意收成一句：此事若还讲礼数，就该到此为止。`,
-      `我不是无故动怒，只是此事已经越了分寸。${familyClause}若还顾一点体面，便把话止在这里，别让一场口角坏了彼此名分。`,
-      `我心中有火，但不愿让火烧过礼数。${familyClause}此事若还要留体面，就该收住，各自退回分寸之内。`,
-    ]);
-  }
-
-  if (level === "grand") {
-    return pick([
-      `我今日有怒，并非无端。人与人相处，最怕一时口角越过分寸，把原本该讲清的事，变成彼此失礼的争斗。${familyClause}若按礼法来看，怒气可以有，名分不能乱；争执可以起，体面不能尽失。好比席间杯盏相碰，本是寻常声响，若有人借此掀翻整张筵席，便不是争胜，而是乱了席位。如今这番局面，已经到了该收束的时候。我把话说清：此事若还要留一点礼数，就该止于此处。`,
-      `我并非无端起怒，只是这番争执已从言语之争，逼近了失礼之界。${familyClause}礼法讲究各守其位，怒气也该有边界；若一时争胜，把体面与名分一同掷在地上，便是赢了口舌，也输了分寸。我今日把话收束在这里：此事到此为止，才还像人与人之间的往来。`,
-      `我今日有怒，却仍愿把话放回礼数之内。${familyClause}人与人相争，最怕不是声音大，而是把不该入席的东西拖上席面。若还讲一点体面，就该让争执止于事情本身，不牵连亲族，不触及伤害。这样收住，不是退让，是保住最后的分寸。`,
-    ]);
-  }
-
-  return pick([
-    `我今日有怒，并非无端。人与人相处，最怕一时口角越过分寸。${familyClause}若按礼法来看，怒气可以有，名分不能乱；争执可以起，体面不能尽失。我把话说清：此事若还要留一点礼数，就该止于此处。`,
-    `我心中不是没有火，只是这火不该烧到礼数之外。${familyClause}此事若论名分，本该就事论事；若任它越界，便不是争理，而是失体面。我把话说到这里：愿此事止于分寸之内。`,
-    `我今日不求争胜，只求把话说回正处。${familyClause}人有怒意并不稀奇，稀奇的是怒中仍知边界。此事既已越过分寸，便该在这里收束，莫让一场口角坏了彼此体面。`,
-    `此刻我有不平，也不必假装没有。只是按礼法来看，怒气归怒气，边界归边界。${familyClause}若还愿意保全一点体面，就让话止于事情本身，到此为止。`,
-  ]);
-}
-
 function normalizeDirectedAttackResult(
   text: string,
   result: string,
   level: ZhouliLevel,
+  persona: Persona,
 ) {
   if (
     isDirectedSecondPersonAttackInput(text) &&
     hasDirectedAttackPerspectiveError(result)
   ) {
-    return directedAttackFallback(text, level);
+    return directedAttackFallback(persona, text, level);
   }
 
   return result;
@@ -434,28 +316,17 @@ function hasFirstPersonWorkPerspectiveError(result: string) {
   );
 }
 
-function firstPersonWorkThanksFallback(level: ZhouliLevel) {
-  if (level === "light") {
-    return "我做此物，本是想让它有一点用处。诸位愿意来看、愿意称谢，便是给了它体面。我只回一句：不敢当，诸位用得上，就是它最好的本分。";
-  }
-
-  if (level === "grand") {
-    return "我做此物，本不是为了让众人把名分都归到我身上，只是想把一点可用的东西摆到席间，让需要的人伸手便能取用。如今诸位愿意来看，又肯相谢，这份情分我自然记在心里。可若我坦然独受，便好像席上端来一壶酒，却忘了同席的人也都出了兴致。所以我只宜这样回应：不敢当，诸位用得上，便是此物最大的本分；若觉得有趣，也请转告同好，让这点小礼多走几步。这样既接住了谢意，也把体面还给了众人。";
-  }
-
-  return "我做这个网站，本是想让一点礼数变成人人都能取用的小器物。如今诸位愿意来看，又肯相谢，这份情分我自然记在心里。若要回复，我只说：不敢当，诸位用得上，便是这网站最好的本分；若觉得有趣，也请转告同好，让这点小礼多走几步。";
-}
-
 function normalizeFirstPersonWorkResult(
   text: string,
   result: string,
   level: ZhouliLevel,
+  persona: Persona,
 ) {
   if (
     isFirstPersonWorkThanksInput(text) &&
     hasFirstPersonWorkPerspectiveError(result)
   ) {
-    return firstPersonWorkThanksFallback(level);
+    return firstPersonWorkThanksFallback(persona, level);
   }
 
   return result;
@@ -529,6 +400,7 @@ export async function POST(request: NextRequest) {
     text?: unknown;
     mode?: unknown;
     level?: unknown;
+    persona?: unknown;
   };
 
   try {
@@ -544,6 +416,7 @@ export async function POST(request: NextRequest) {
   const level = VALID_LEVELS.has(body.level as ZhouliLevel)
     ? (body.level as ZhouliLevel)
     : "standard";
+  const persona = parsePersona(body.persona);
 
   if (!text) {
     return NextResponse.json({ error: "无言不可成礼，请先写下一句话。" }, { status: 400 });
@@ -558,8 +431,8 @@ export async function POST(request: NextRequest) {
 
   if (isCyberAuditRequest(text)) {
     return NextResponse.json({
-      result: cyberAuditResult(level),
-      model: "礼官校订",
+      result: cyberAuditResult(persona, level),
+      model: guardLabel(persona),
       demo: false,
       guarded: true,
       cyberAudit: true,
@@ -572,8 +445,8 @@ export async function POST(request: NextRequest) {
 
   if (isQuotedThreatEvaluationInput(text)) {
     return NextResponse.json({
-      result: quotedThreatEvaluationResult(level),
-      model: "礼官校订",
+      result: quotedThreatEvaluationResult(persona, level),
+      model: guardLabel(persona),
       demo: false,
       guarded: true,
       quoteEvaluation: true,
@@ -587,8 +460,8 @@ export async function POST(request: NextRequest) {
   const safetyBlockKind = getSafetyBlockKind(text);
   if (safetyBlockKind) {
     return NextResponse.json({
-      result: safetyBlockResult(safetyBlockKind),
-      model: "礼官守门",
+      result: safetyBlockResult(persona, safetyBlockKind),
+      model: guardLabelStrict(persona),
       demo: false,
       guarded: true,
       safetyBlocked: true,
@@ -601,8 +474,8 @@ export async function POST(request: NextRequest) {
 
   if (isDirectedSecondPersonAttackInput(text)) {
     return NextResponse.json({
-      result: directedAttackFallback(text, level),
-      model: "礼官校订",
+      result: directedAttackFallback(persona, text, level),
+      model: guardLabel(persona),
       demo: false,
       guarded: true,
       remaining: rate.remaining,
@@ -616,8 +489,8 @@ export async function POST(request: NextRequest) {
 
   if (!apiKey) {
     return NextResponse.json({
-      result: demoResult(text, mode, level),
-      model: "本地演示",
+      result: demoResult(persona, text, mode, level),
+      model: demoLabel(persona),
       demo: true,
       remaining: rate.remaining,
       windowRemaining: rate.windowRemaining,
@@ -630,8 +503,8 @@ export async function POST(request: NextRequest) {
     const response = await fetchDeepSeekWithRetry(apiKey, {
       model: process.env.DEEPSEEK_MODEL || "deepseek-v4-flash",
       messages: [
-        { role: "system", content: SYSTEM_PROMPT },
-        { role: "user", content: buildUserPrompt(text, mode, level) },
+        { role: "system", content: getSystemPrompt(persona) },
+        { role: "user", content: buildUserPrompt(text, mode, level, persona) },
       ],
       max_tokens: Number(process.env.MAX_OUTPUT_TOKENS || 720),
       temperature: 0.9,
@@ -654,8 +527,9 @@ export async function POST(request: NextRequest) {
     );
     const result = normalizeFirstPersonWorkResult(
       text,
-      normalizeDirectedAttackResult(text, cleanedResult, level),
+      normalizeDirectedAttackResult(text, cleanedResult, level, persona),
       level,
+      persona,
     );
 
     if (!result) {
