+++ app/layout.tsx
@@ -2,12 +2,12 @@ import type { Metadata } from "next";
 import "./globals.css";
 
 export const metadata: Metadata = {
-  title: "合乎周礼｜把寻常的话，说得有礼有据",
+  title: "儒释道｜把寻常的话，说得有礼有据",
   description:
     "输入一句寻常话，把它翻译成大周礼时代的白话翻译腔。现代白话为骨，古代典故为证。",
-  keywords: ["大周礼时代", "合乎周礼", "AI翻译", "网络梗", "DeepSeek"],
+  keywords: ["大周礼时代", "儒释道", "AI翻译", "网络梗", "DeepSeek"],
   openGraph: {
-    title: "合乎周礼",
+    title: "儒释道",
     description: "把寻常的话，说得有礼有据。",
     type: "website",
     locale: "zh_CN",
