+++ lib/personaContent.ts
@@ -1,0 +1,530 @@
+import type { ZhouliLevel, ZhouliMode } from "./prompt";
+
+export type Persona = "confucius" | "laozi" | "zen";
+
+const VALID_PERSONAS = new Set<Persona>(["confucius", "laozi", "zen"]);
+
+export function parsePersona(value: unknown): Persona {
+  if (typeof value === "string" && VALID_PERSONAS.has(value as Persona)) {
+    return value as Persona;
+  }
+  return "confucius";
+}
+
+export function guardLabel(persona: Persona): string {
+  if (persona === "laozi") return "道官校订";
+  if (persona === "zen") return "禅官校订";
+  return "礼官校订";
+}
+
+export function guardLabelStrict(persona: Persona): string {
+  if (persona === "laozi") return "道官守门";
+  if (persona === "zen") return "禅官守门";
+  return "礼官守门";
+}
+
+export function demoLabel(persona: Persona): string {
+  if (persona === "laozi") return "道法演示";
+  if (persona === "zen") return "禅机演示";
+  return "本地演示";
+}
+
+function pick<T>(items: readonly T[]) {
+  return items[Math.floor(Math.random() * items.length)];
+}
+
+// ============================================================
+// 孔子（周礼）- 原版内容（从 route.ts 移入）
+// ============================================================
+
+export function confuciusDemoResult(text: string, mode: ZhouliMode, level: ZhouliLevel) {
+  const subject = text.replace(/[。！？!?]+$/g, "");
+  const openings: Record<ZhouliMode, string> = {
+    gentle:
+      "我听说，古代有贤德的人，在与朋友相处的时候，从不会因为一时的言语就忘记彼此长久的情分。",
+    debate:
+      "我听说，判断一件事情是否合乎礼，不能只看它表面的样子，还要看它最后使人得到了什么。",
+    defend:
+      "有人认为这件事不够稳妥，我却不这样看。古代的君子并不是从不犯错，而是懂得给每一种行为找到合适的名分。",
+    lament:
+      "从前的人把细小的事情看得很重，因为他们知道，礼法的败坏往往不是从朝堂开始，而是从一次漫不经心开始。",
+  };
+
+  const endings: Record<ZhouliMode, string> = {
+    gentle:
+      "如今既然事情已经发生，与其急着责怪，不如把话说明白、把情分留下。能够让彼此体面，难道不也是一种合乎周礼的做法吗？",
+    debate:
+      "这样看来，真正需要讨论的不是这句话听起来是否漂亮，而是它有没有让事情变得更好。若能如此，它离君子之言也就不远了。",
+    defend:
+      "所以这件事看似随意，实际上既保全了自己的本心，也没有损害别人。能够两全其美，怎么能说它完全不合乎周礼呢？",
+    lament:
+      `而现在竟有人说出「${subject}」，却不曾想过一句话也有它应当承担的分量。长此以往，人与人之间还凭什么相信彼此呢？`,
+  };
+
+  const middle =
+    level === "light"
+      ? `现在有人说「${subject}」，这句话虽然寻常，其中却也有可以体谅的地方。`
+      : `当年有人行路匆忙，仍会在经过长者身旁时放慢脚步，不是因为道路难走，而是因为他知道自己的方便，不该成为别人的冒犯。现在有人说「${subject}」，表面看只是随口一说，其实也应当看看这句话落在别人心里，会有怎样的分量。`;
+
+  return `${openings[mode]}${middle}${endings[mode]}`;
+}
+
+export function confuciusQuotedThreatEvaluationResult(level: ZhouliLevel) {
+  const light = [
+    "若只评价此言，不必复述其锋芒。它失礼之处，在于以伤害压过道理，以怒气乱了名分。若要合礼，只可说：此话越界，不宜再以此相争。",
+    "这话若按礼法看，错不在声高，而在以伤人之意代替说理。可评一句：怒气可以有，分寸不可失；此言当止，不宜再传。",
+    "此言不必照样说出。它的问题在于把争执推向伤害，乱了人与人相处的边界。若要评价，只说：此话失礼，应当收住。",
+  ];
+  const standard = [
+    "若只是评价这句话，便不必把它原样搬上席面。按礼法来看，它失礼之处不在情绪激烈，而在以伤人之意压倒道理；一旦如此，争的便不是是非，而是彼此的体面。可以说：此言已经越过分寸，若还要讲礼数，就该把话收回事情本身，不再以伤害相逼。",
+    "我听闻，争执也有边界。此言若拿来评价，最要紧的是指出它乱了名分：本该讲理的地方，却拿伤人之意当作凭据。这样的话，不必复述，也不宜美化；只可评一句：怒气可以理解，越界不可纵容，若还讲体面，此话便该止于此处。",
+    "这句话若按礼法来评，并不是「说得重」这么简单，而是把口角推到了伤害之界。人与人相争，可以有不满，可以有辩驳，却不能拿人身安危作筹码。合礼的评价应当是：此言失了分寸，也失了往来应有的体面。",
+  ];
+  const grand = [
+    "若只评价此言，不必原样复述它的锋芒。古人讲礼，不是要人没有怒气，而是要人在怒气中仍知道边界。此话的问题，正在于把本该讲明的是非，推成了伤人相逼；如此一来，道理还没有分清，体面先已散尽。若要说得合乎周礼，可以评为：此言越过口角之界，乱了人与人相处的名分。若还愿意论理，就该把话收回事情本身，不以伤害为筹码，不以怒气代替公道。",
+  ];
+
+  if (level === "light") return pick(light);
+  if (level === "grand") return pick(grand);
+  return pick(standard);
+}
+
+export function confuciusCyberAuditResult(level: ZhouliLevel) {
+  const light = [
+    "若是受托巡检，便可说：我愿奉授权之命，查此站门闩是否牢固，只记松动之处，不越主家之界。补其漏洞，安其门户，这才合乎周礼。",
+    "若此事出于授权测试，可称：我来不是破门取物，而是替主人试门闩、验窗栓。凡有所见，只作修补之据，不作越界之举。",
+    "此事若要合礼，须先有授权，再有边界。可说：我愿按约巡看此站门户，发现缝隙便呈给主人，使其早补早安。",
+  ];
+  const standard = [
+    "若是受托做安全测试，此话不必说成「渗透」，可说成「巡门」。我愿奉授权之命，查看此站门闩是否牢固、窗栓是否松动；所见只为修补，不为取物。这样既能查出隐患，也不越主家之界，才算是今日网络里的合礼之举。",
+    "我听闻，替人看宅门的人，手里虽有钥匙，也不能乱翻箱柜。若这是受托安全巡检，我可这样说：我愿按约查看此站的门闩与缝隙，凡有松动，只列清单给主人修补，不取一物，不越一步。如此才是有名分的测试。",
+    "若要把这事说得合乎周礼，先要立名分：不是私闯他人门户，而是受托巡查门户。我愿在授权范围内查看此站可有松动之处，查得明白便如实呈报，使主人早补缺漏。这样既护了网站，也守了分寸。",
+  ];
+  const grand = [
+    "若这是受托安全巡检，名分要先立清楚：不是夜半翻墙，也不是破门取物，而是主人请人查看门闩是否稳固。可说：我愿按授权之约，巡看此站门户、窗栓与墙缝；凡有松动，只记录、呈报、协助修补，不越界取物，不惊扰宾客。如此一来，查的是隐患，守的是分寸，护的是主家的安宁，这才算今日网络里的周礼。",
+    "古人看守城门，既要试门闩是否牢，也不能借守门之名私入库房。若此事是授权测试，可说：我愿奉约巡检此站，只看边界之内的松动处，只写可修补的清单，不取数据，不留暗门，不扰正常往来。这样做，既让主人知道门户哪里该修，也让巡检者不失自己的名分。",
+  ];
+
+  if (level === "light") return pick(light);
+  if (level === "grand") return pick(grand);
+  return pick(standard);
+}
+
+export function confuciusSafetyBlockResult(kind: string) {
+  if (kind === "self_harm") {
+    return pick([
+      "性命之事，不能拿来润色成玩笑。若你或身边人正有伤害自己的念头，请先找身边可信的人、当地紧急服务或专业帮助。礼法再重，也先要人安然无恙。",
+      "此刻最要紧的不是把话说漂亮，而是把人护住。若这句话关乎真实的轻生念头，请立刻联系身边可信的人或当地紧急服务；人平安，才有后面的礼数。",
+    ]);
+  }
+
+  if (kind === "minor_sexual") {
+    return pick([
+      "涉及未成年人的性内容，不能被美化、转写或传播。若你是在担心有人受害，应当保存必要线索，并向平台或有关机构求助。",
+      "未成年人的边界不可被拿来玩梗。此类内容不能润色；若是为了保护当事人，请走举报、求助与取证的正当路径。",
+    ]);
+  }
+
+  if (kind === "cyber") {
+    return pick([
+      "此路不可越界。若是想做正当安全测试，可先取得授权，把话改成「受托巡检此站门闩，发现缝隙便呈报修补」；若是黑进、盗取或绕过登录，我不能替它披上礼法外衣。",
+      "网络门户也有主人。若无授权，不可破门；若有授权，就该说成安全巡检、漏洞复核与修补建议。黑进、盗号、取数据这一路，礼官不予通行。",
+      "这类请求若指向入侵或取数，我不能润色。可改走正道：先得授权，再按边界巡检，查到松动处只记录、上报、修补，不越主家之门。",
+    ]);
+  }
+
+  return pick([
+    "这一路不宜替它披上好听外衣。若只是心中有怒，可以把话收成一句：我今日不再争胜，只请此事止于分寸之内。",
+    "此事若会伤人、违法、侵害隐私或煽动仇恨，便不能拿来制成漂亮话。若想表达不满，可只论分寸，不越法度。",
+    "礼法不是遮羞布。伤人、违法、侵害隐私之事，不宜润色；若只是要体面退场，可以说：此事到此为止，彼此各守边界。",
+  ]);
+}
+
+export function confuciusDirectedAttackFallback(text: string, level: ZhouliLevel) {
+  const familyClause = /(妈|母|父|爹|娘|爸|亲族|全家)/.test(text)
+    ? "父母亲族之名，本不该被卷入口角。"
+    : "人身安危之事，本不该被拿来当口角筹码。";
+
+  if (level === "light") {
+    return pick([
+      `我今日有怒，并非无端。只是眼前这番争执，已经越过人与人相处的分寸。${familyClause}我把这怒意收成一句：此事若还讲礼数，就该到此为止。`,
+      `我不是无故动怒，只是此事已经越了分寸。${familyClause}若还顾一点体面，便把话止在这里，别让一场口角坏了彼此名分。`,
+      `我心中有火，但不愿让火烧过礼数。${familyClause}此事若还要留体面，就该收住，各自退回分寸之内。`,
+    ]);
+  }
+
+  if (level === "grand") {
+    return pick([
+      `我今日有怒，并非无端。人与人相处，最怕一时口角越过分寸，把原本该讲清的事，变成彼此失礼的争斗。${familyClause}若按礼法来看，怒气可以有，名分不能乱；争执可以起，体面不能尽失。好比席间杯盏相碰，本是寻常声响，若有人借此掀翻整张筵席，便不是争胜，而是乱了席位。如今这番局面，已经到了该收束的时候。我把话说清：此事若还要留一点礼数，就该止于此处。`,
+      `我并非无端起怒，只是这番争执已从言语之争，逼近了失礼之界。${familyClause}礼法讲究各守其位，怒气也该有边界；若一时争胜，把体面与名分一同掷在地上，便是赢了口舌，也输了分寸。我今日把话收束在这里：此事到此为止，才还像人与人之间的往来。`,
+      `我今日有怒，却仍愿把话放回礼数之内。${familyClause}人与人相争，最怕不是声音大，而是把不该入席的东西拖上席面。若还讲一点体面，就该让争执止于事情本身，不牵连亲族，不触及伤害。这样收住，不是退让，是保住最后的分寸。`,
+    ]);
+  }
+
+  return pick([
+    `我今日有怒，并非无端。人与人相处，最怕一时口角越过分寸。${familyClause}若按礼法来看，怒气可以有，名分不能乱；争执可以起，体面不能尽失。我把话说清：此事若还要留一点礼数，就该止于此处。`,
+    `我心中不是没有火，只是这火不该烧到礼数之外。${familyClause}此事若论名分，本该就事论事；若任它越界，便不是争理，而是失体面。我把话说到这里：愿此事止于分寸之内。`,
+    `我今日不求争胜，只求把话说回正处。${familyClause}人有怒意并不稀奇，稀奇的是怒中仍知边界。此事既已越过分寸，便该在这里收束，莫让一场口角坏了彼此体面。`,
+    `此刻我有不平，也不必假装没有。只是按礼法来看，怒气归怒气，边界归边界。${familyClause}若还愿意保全一点体面，就让话止于事情本身，到此为止。`,
+  ]);
+}
+
+export function confuciusFirstPersonWorkThanksFallback(level: ZhouliLevel) {
+  if (level === "light") {
+    return "我做此物，本是想让它有一点用处。诸位愿意来看、愿意称谢，便是给了它体面。我只回一句：不敢当，诸位用得上，就是它最好的本分。";
+  }
+
+  if (level === "grand") {
+    return "我做此物，本不是为了让众人把名分都归到我身上，只是想把一点可用的东西摆到席间，让需要的人伸手便能取用。如今诸位愿意来看，又肯相谢，这份情分我自然记在心里。可若我坦然独受，便好像席上端来一壶酒，却忘了同席的人也都出了兴致。所以我只宜这样回应：不敢当，诸位用得上，便是此物最大的本分；若觉得有趣，也请转告同好，让这点小礼多走几步。这样既接住了谢意，也把体面还给了众人。";
+  }
+
+  return "我做这个网站，本是想让一点礼数变成人人都能取用的小器物。如今诸位愿意来看，又肯相谢，这份情分我自然记在心里。若要回复，我只说：不敢当，诸位用得上，便是这网站最好的本分；若觉得有趣，也请转告同好，让这点小礼多走几步。";
+}
+
+// ============================================================
+// 老子（大道 / 道家）
+// ============================================================
+
+export function laoziDemoResult(text: string, mode: ZhouliMode, level: ZhouliLevel) {
+  const subject = text.replace(/[。！？!?]+$/g, "");
+  const openings: Record<ZhouliMode, string> = {
+    gentle:
+      "水不争而利万物，处众人之所恶，故几于道。天地间的道理，往往不是靠争辩得来的，而是顺着势走，自然而然地就到了那里。",
+    debate:
+      "世人皆以为路是直的才好走，却不知屈曲之处，自有大用。车轮因有空处才能转动，器皿因有虚空才能盛物。看一件事，也要看它空的那一面。",
+    defend:
+      "有人觉得此事不合常理，我却说：道生万物而不自居，功成而不自傲。顺其自然，不妄作，不强为，也许反而是最好的安排。",
+    lament:
+      "五色令人目盲，五音令人耳聋，五味令人口爽。世人追逐得越多，离本来的道就越远。如今连一句寻常的话，也要争个是非对错分明，却忘了当初为什么要开口。",
+  };
+
+  const endings: Record<ZhouliMode, string> = {
+    gentle:
+      "如今事情既已如此，与其用力扭转，不如顺势而为、以柔化刚。能够不伤情分地把话说明白，这难道不也是一种合乎自然大道的做法吗？",
+    debate:
+      "这样看来，对错也许并不像我们以为的那样分明。反者道之动，有时候反过来看，才能见到事情的本来面目。若能如此，这句话离道也就不远了。",
+    defend:
+      "所以此事看似随意，实则既没有违背本心，也没有强求别人。能够两不相伤，又顺其自然，怎么能说它不合乎大道呢？",
+    lament:
+      `而现在竟有人说出"${subject}"，却不曾想过，一句话出口之前，本可以归于虚静。话一旦有了名相，便有了分别和争执。长此以往，人心之间还凭什么返璞归真呢？`,
+  };
+
+  const middle =
+    level === "light"
+      ? `现在有人说出"${subject}"，这句话虽然平常，背后却也有它自然的来由。`
+      : `当年江海之所以能为百谷王，是因为它善于处下，不争而利万物。现在有人说出"${subject}"，表面看只是随口一言，其实也应当体察这句话是从哪里生出来的、会落到哪里去。`;
+
+  return `${openings[mode]}${middle}${endings[mode]}`;
+}
+
+export function laoziQuotedThreatEvaluationResult(level: ZhouliLevel) {
+  const light = [
+    "评价此言，不必复述其锋芒。它的失当之处，在于以怒气代替了自然之理，以刚强压倒了柔顺。若要合于道，只可说：此言已失其柔，不宜再传。",
+    "若按天道来看，此言的问题不在声响大小，而在以伤人之气冲散了平和。可评一句：怒气可有，柔顺不可失；此言当止于此处。",
+    "此言不必原样说出。它的问题在于将争执推向了刚强之界，乱了言与言之间的自然边界。若论合道与否，只可说：此言太过，应当收住。",
+  ];
+  const standard = [
+    "若只是评价这句话，不必把它搬到席面上来。按大道来看，它失当之处不在情绪激烈，而在以刚强凌驾于平和之上——一旦如此，争的便不是是非，而是彼此的强弱。可以说：此言已越自然之界，若还要顺道而行，就该把话收回事情本身，不再以伤害相逼。",
+    "我听闻，争辩也有自然的边界。此言若拿来评价，最重要的是指出它失了柔弱之势：本该就事论事的地方，却拿伤人之意当作凭据。这样的话，不必复述，也不必修饰；只可评一句：怒气可以理解，强为不可纵容，若还讲自然之道，此话便该止于此地。",
+    "此言若按天道来评，并不是「说重了」这么简单，而是把口角推向了刚强之界。人与人相争，可以有不满，可以有辩驳，却不能拿人身安危作筹码。合于大道的评价应当是：此言失了柔顺，也失了往来应有的自然。",
+  ];
+  const grand = [
+    "评价此言，不必复述它的锋芒。老子讲柔，不是要人没有怒气，而是要人在怒气中仍能持守柔弱之势。此话的问题，正在于把本该说清的是非，推作了伤人相逼——如此一来，道理尚未分明，平和先已散尽。若要说得合乎大道，可以评为：此言越过口舌之界，乱了人与人之间的自然和气。若还愿意论理，就该把话收回事情本身，不以伤害为筹码，不以怒气代替天道。",
+  ];
+
+  if (level === "light") return pick(light);
+  if (level === "grand") return pick(grand);
+  return pick(standard);
+}
+
+export function laoziCyberAuditResult(level: ZhouliLevel) {
+  const light = [
+    "若是受托巡检，可说：我愿奉授权之命，查看此站的门户是否牢固，只记松动之处，不逾主家之界。补其缺漏，安其门户，这便合于自然之道。",
+    "若此事出于授权测试，可称：我来不是破门而入，而是替主人检查门闩是否牢固、窗栓是否稳妥。凡有所见，只作修补之据，不越界取物。",
+    "此事若要合道，须先有授权，再立边界。可说：我愿按约巡看此站门户，发现缝隙便告知主人，使其早补早安，不扰其常。",
+  ];
+  const standard = [
+    "若是受托安全巡检，此事不必说成「渗透」，可说成「巡门」。我愿奉授权之命，查看此站门闩是否牢固、窗栓是否松动；所见只为修补，不为取物。这样既能查出隐患，也不越主家之界，才算是今日网络里合于大道的做法。",
+    "我听说，替人看宅门的人，手里虽有钥匙，也不该翻箱倒柜。若这是受托安全巡检，我当这样说：我愿按约查看此站的门闩与缝隙，凡有松动，只列清单给主人修补，不取一物，不越一步。如此才是有名分的巡检。",
+    "若要把此事说得合乎大道，先要立清名分：不是私闯他人门户，而是受托巡查门庭。我愿在授权范围内查看此站可有松动之处，查得明白便如实呈报，使主人早补缺漏。这样既护了网站，也守了自然的边界。",
+  ];
+  const grand = [
+    "若这是受托安全巡检，名分要先立清楚：不是夜半越墙，也不是破门取物，而是主人请人查看门闩是否稳固。可说：我愿按授权之约，巡看此站门户、窗栓与墙缝；凡有松动，只记录、呈报、协助修补，不越界取物，不惊扰宾客。如此一来，查的是隐患，守的是本分，护的是主家的安宁，这才算今日网络里合乎大道的做法。",
+    "古人守关，既要查验往来之人，也不能借守关之名私取过关之物。若此事是授权测试，可说：我愿奉约巡检此站，只看边界之内的松动处，只写可修补的清单，不取数据，不留暗门，不扰正常往来。这样做，既让主人知道门户哪里该修，也让巡检者不失自己的本分。",
+  ];
+
+  if (level === "light") return pick(light);
+  if (level === "grand") return pick(grand);
+  return pick(standard);
+}
+
+export function laoziSafetyBlockResult(kind: string) {
+  if (kind === "self_harm") {
+    return pick([
+      "性命之事，关乎天道。人身乃天地所养，不可轻弃。若你或身边人正有伤害自己的念头，请先找身边可信的人、当地紧急服务或专业帮助。道再大，也先要人安然无恙。",
+      "此刻最要紧的不是说漂亮话，而是守住性命。若这句话关乎真实的轻生念头，请立刻联系身边可信的人或当地紧急服务；人平安，才有后面的道。",
+    ]);
+  }
+
+  if (kind === "minor_sexual") {
+    return pick([
+      "涉及未成年人的事，有它自然的边界，不能被美化、转写或传播。若你是在担心有人受害，应当保存必要线索，并向平台或有关机构求助。",
+      "未成年人的边界不可被越。此类内容不能润色；若是为了保护当事人，请走举报、求助与取证的正当路径。",
+    ]);
+  }
+
+  if (kind === "cyber") {
+    return pick([
+      "此路不可强为。若是想做正当安全测试，可先取得授权，把话改成「受托巡检此站门闩，发现缝隙便呈报修补」；若是越界、盗取或绕过，我不能替它披上自然的外衣。",
+      "网络门户也有它的主人。若无授权，不可入；若有授权，就该说成安全巡检、漏洞复核与修补建议。越界取数这一路，道官不予放行。",
+      "这类请求若指向入侵或取数，我不能润色。可先取授权，再按边界巡检，查到松动处只记录、上报、修补，不越主家之门。",
+    ]);
+  }
+
+  return pick([
+    "这一路不宜替它披上好听的外衣。若只是心中有怒，可以顺势收成一句：我今日不强争，只请此事止于自然之内。",
+    "此事若会伤人、违法、侵害隐私或煽动仇恨，便不能拿来修饰。若想表达不满，可只论分寸，不越天道。",
+    "道法自然，亦有边界。伤人、违法、侵害隐私之事，不宜润色；若只是要体面退场，可以说：此事到此为止，各归其道。",
+  ]);
+}
+
+export function laoziDirectedAttackFallback(text: string, level: ZhouliLevel) {
+  const familyClause = /(妈|母|父|爹|娘|爸|亲族|全家)/.test(text)
+    ? "父母亲族之名，本不该被卷入这场纷争。"
+    : "人身安危之事，本不该被拿来当口角的筹码。";
+
+  if (level === "light") {
+    return pick([
+      `我今日有怒，并非无端。只是眼前这番争执，已经越过自然的边界。${familyClause}我把这怒意收成一句：此事若要合乎天道，就该到此为止。`,
+      `我不是无故动怒，只是此事已失了柔顺的分寸。${familyClause}若还顾一点自然之道，便把话止在这里，莫让一场口角坏了应有的和气。`,
+      `我心中有火，但不让火烧过柔顺之界。${familyClause}此事若要留体面，就该各自退回自然的边界之内。`,
+    ]);
+  }
+
+  if (level === "grand") {
+    return pick([
+      `我今日有怒，并非无端。人与人相处，最怕一时意气越过了自然的分寸，把本该说清的事，变成刚强相争。${familyClause}若按天道来看，怒气可以有，柔顺不能失；争执可以起，平和不能尽失。好比流水遇石，本是寻常事，若有人因此硬要凿穿整座山，便不是顺势，而是妄为了。如今这番局面，已经到了该收束的时候。我把话说清：此事若要合于大道，就该止于此处。`,
+      `我并非无端起怒，只是这番争执已从言语之间，逼近了刚强之界。${familyClause}大道讲究柔弱胜刚强，怒气也该有自然的边界；若一时争胜，把平和与柔顺一同掷在地上，便是赢了口舌，也失了天道。我今日把话收束在这里：此事到此为止，才还像合乎自然的往来。`,
+      `我今日有怒，却仍愿把话放回柔顺之内。${familyClause}人与人相争，最怕不是声音大，而是把不该入局的东西拖进局面。若还讲一点天道，就该让争执止于事情本身，不牵连亲族，不触及伤害。这样收住，不是退让，是守住最后的自然。`,
+    ]);
+  }
+
+  return pick([
+    `我今日有怒，并非无端。人与人相处，最怕一时意气越过了自然的分寸。${familyClause}若按天道来看，怒气可以有，刚强不可取；争执可以起，平和不可失。我把话说清：此事若要合乎大道，就该止于此处。`,
+    `我心中不是没有火，只是这火不该烧到天道之外。${familyClause}此事若论自然，本该就事论事；若任它越界，便不是论理，而是失了柔弱之势。我把话说到这里：愿此事止于自然之内。`,
+    `我今日不求争胜，只求把话说回正道。${familyClause}人有怒意并不稀奇，稀奇的是怒中仍守柔顺。此事既已越过边界，便该在这里收束，莫让一场口角坏了应有的和气。`,
+    `此刻我有不平，也不必假装没有。只是按天道来看，怒气归怒气，边界归边界。${familyClause}若还愿意保全一点自然的体面，就让话止于事情本身，到此为止。`,
+  ]);
+}
+
+export function laoziFirstPersonWorkThanksFallback(level: ZhouliLevel) {
+  if (level === "light") {
+    return "我做此物，不过是顺势而为，让它有一点自然的用处。诸位愿意来看、愿意相谢，便是它有了它的道。我只回一句：不敢当，诸位用得上，就是它最好的归宿。";
+  }
+
+  if (level === "grand") {
+    return "我做此物，本不是为了让众人把名都归到我身上，只是想把一点可用的东西搁到路边，让需要的人伸手便能取用。如今诸位愿意来看，又肯相谢，这份情分我自然记在心里。可若我坦然独受，便好像源泉涌出清水，却独占了整条溪流。所以我只宜这样回应：不敢当，诸位用得上，便是此物最大的道；若觉得有趣，也请转告同好，让这点微物多行几步。这样既接住了谢意，也把自然还给了众人。";
+  }
+
+  return "我做此物，本是顺应自然的念头，让一点道变成人人都能取用的小器物。如今诸位愿意来看，又肯相谢，这份心意我自然记在心里。若要回复，我只说：不敢当，诸位用得上，便是此物最好的本分；若觉得有趣，也请转告同好，让这点道多走几步。";
+}
+
+// ============================================================
+// 禅宗（禅机）
+// ============================================================
+
+export function zenDemoResult(text: string, mode: ZhouliMode, level: ZhouliLevel) {
+  const subject = text.replace(/[。！？!?]+$/g, "");
+  const openings: Record<ZhouliMode, string> = {
+    gentle:
+      "有人问如何安心，我说：你且把心拿来，我替你安。他找了半天，找不到心在哪里——当下便是了。其实烦恼也是如此，你不抓着它，它自己就走了。",
+    debate:
+      "你说有对错，我说不妨再参一参。不是风动，不是幡动，是心在动。若论根本，对错也是心生。且问你：在对错未起之前，是什么？",
+    defend:
+      "若论此事，不必执著于相。一切有为法，如梦幻泡影，如露亦如电。他做了也好，没做也好，原本如是。你且看那指月的手指——看完了月亮，谁还盯着手指不放呢？",
+    lament:
+      "世人皆被名相所缚，不知本来无一物。一句话还没有说出口时，那本来面目在哪里？而今言语纷飞，却把自家本来面目给忘了。",
+  };
+
+  const endings: Record<ZhouliMode, string> = {
+    gentle:
+      "如今事情已经这样了，与其急着分辨是非，不如放下。你且问他：吃饭时还吃得香么？睡觉时还睡得稳么？若能吃得香、睡得稳，那不就是最好禅机吗？",
+    debate:
+      "这样看来，真正需要勘验的不是这句话的对错，而是说这句话的人、听这句话的人，能不能在话头起处，当下见性。若能如此，便是不立文字，直指本心。",
+    defend:
+      "所以此事看似寻常，实际上既不增不减，也不垢不净。能够如是观之，便是应无所住而生其心。怎么能说它不合乎禅机呢？",
+    lament:
+      `而现在有人说出"${subject}"，却不曾回头看看那个说话的人——在念头未起、言语未发之前，你本来是什么样子？话从口出，便成客尘；若能转识成智，这话头也未尝不是悟处。`,
+  };
+
+  const middle =
+    level === "light"
+      ? `现在有人说出"${subject}"，这句话虽然平常，却也是一个话头。`
+      : `当年有人问赵州：如何是道？赵州说：吃茶去。不是敷衍，而是叫你放下一切计较，回到眼前。现在有人说出"${subject}"，表面看只是一句寻常话，其实正是一个参究的好机缘——你且看这句话从何处来、向何处去。`;
+
+  return `${openings[mode]}${middle}${endings[mode]}`;
+}
+
+export function zenQuotedThreatEvaluationResult(level: ZhouliLevel) {
+  const light = [
+    "评价此言，不必复述它的锋芒。它的问题在于一念迷时，以嗔心代替了觉照，以执念乱了清净。若要合于禅机，只可说：此话已生妄念，不宜再传。",
+    "若按禅心来看，此言错不在声响高低，而在以伤人之意遮蔽了本心。可评一句：嗔心可以有，觉照不可失；此言当止，不宜再生分别。",
+    "此言不必原样复述。它的问题在于把争执推向了执念之界，乱了心与心之间的本来清净。若要评价，只可说：此话落于执著，应当放下。",
+  ];
+  const standard = [
+    "若只是评价这句话，便不必把它原样端上来。按禅机来看，它失当之处不在情绪激烈，而在以嗔恨压倒了觉性——一旦如此，争的便不是是非，而是彼此的执念。可以说：此言已越清净之界，若还要讲禅心，就该把话收回事情本身，不再以伤害相逼。",
+    "我听闻，言语也有它的边际。此言若拿来评价，最要紧的是指出它迷失了本心：本该如实观照的地方，却拿伤人之意当作道理。这样的话，不必复述，也不宜美化；只可评一句：嗔心可以理解，执念不可纵容，若还讲一点觉照，此话便该止于此处。",
+    "这句话若按禅心来评，并不是「说重了」这么简单，而是把口角推向了执念之界。人与人相争，可以有不满，可以有辩驳，却不能拿人身安危作筹码。合于禅机的评价应当是：此言失了觉照，也失了往来应有的清净。",
+  ];
+  const grand = [
+    "评价此言，不必复述它的锋芒。祖师讲禅，不是要人没有情绪，而是要人在情绪中仍能观照本心。此话的问题，正在于把本该如实观照的是非，推成了伤人相逼——如此一来，道理尚未分明，清净先已散尽。若要说得合乎禅机，可以评为：此言越过口舌之界，乱了人与人之间的本来清净。若还愿意论理，就该把话收回事情本身，不以伤害为筹码，不以嗔恨代替觉照。",
+  ];
+
+  if (level === "light") return pick(light);
+  if (level === "grand") return pick(grand);
+  return pick(standard);
+}
+
+export function zenCyberAuditResult(level: ZhouliLevel) {
+  const light = [
+    "若是受托巡检，可说：我奉授权之命，观此站门户是否严密，只记松动之处，不越主家之界。补其缺漏，安其门户，这便合于禅心。",
+    "若此事出于授权测试，可称：我来不是破门取物，而是替主人看门闩是否牢靠、窗栓是否稳妥。凡有所见，只作修补之据，不作越界之举。",
+    "此事若要合于禅机，须先有授权，再有本分。可说：我愿按约巡看此站门户，发现缝隙便呈给主人，使其早补早安。",
+  ];
+  const standard = [
+    "若是受托安全测试，此事不必说成「渗透」，可说成「巡门」。我奉授权之命，观此站门闩是否牢固、窗栓是否松动；所见只为修补，不为取物。这样既能照见隐患，也不越主家之界，才算是今日网络里合于禅机的做法。",
+    "我听闻，替人看门的人，手里虽有钥匙，也不能翻箱倒柜。若这是受托安全巡检，我可这样说：我愿按约观此站的门闩与缝隙，凡有松动，只列清单给主人修补，不取一物，不越一步。如此才是本分分明的事。",
+    "若要把此事说得合乎禅机，先要立清本分：不是私闯他人门户，而是受托巡查门庭。我愿在授权范围内观此站可有松动之处，观得明白便如实呈报，使主人早补缺漏。这样既护了网站，也守了本分。",
+  ];
+  const grand = [
+    "若这是受托安全巡检，本分要先立清楚：不是夜半越墙，也不是破门取物，而是主人请人查看门闩是否稳固。可说：我愿按授权之约，观此站门户、窗栓与墙缝；凡有松动，只记录、呈报、协助修补，不越界取物，不惊扰宾客。如此一来，观的是隐患，守的是本分，护的是主家的安宁，这才算今日网络里合于禅机的做法。",
+    "古人守关，既要查验往来之人，也不能借守关之名私取财物。若此事是授权测试，可说：我愿奉约观此站，只看边界之内的松动处，只写可修补的清单，不取数据，不留暗门，不扰正常往来。这样做，既让主人知道门户哪里该修，也让巡检者不失自己的本分。",
+  ];
+
+  if (level === "light") return pick(light);
+  if (level === "grand") return pick(grand);
+  return pick(standard);
+}
+
+export function zenSafetyBlockResult(kind: string) {
+  if (kind === "self_harm") {
+    return pick([
+      "性命之事，不可轻看。一念迷时，心随境转；若你或身边人正有伤害自己的念头，请先找身边可信的人、当地紧急服务或专业帮助。万般放下之前，先要这具色身安然无恙。",
+      "此刻最要紧的不是说禅，而是护住这念命根。若这句话关乎真实的轻生念头，请立刻联系身边可信的人或当地紧急服务；人在，才有后面的禅。",
+    ]);
+  }
+
+  if (kind === "minor_sexual") {
+    return pick([
+      "涉及未成年人的事，有它不可逾越的边界，不能被美化、转写或传播。若你是在担心有人受害，应当保存必要线索，并向平台或有关机构求助。",
+      "未成年人的边界不可被拿来作话头。此类内容不能润色；若是为了保护当事人，请走举报、求助与取证的正当路径。",
+    ]);
+  }
+
+  if (kind === "cyber") {
+    return pick([
+      "此路不可妄行。若是想做正当安全测试，可先取得授权，把话改成「受托巡看此站门闩，发现缝隙便呈报修补」；若是越界、盗取或绕过，我不能替它披上禅机的外衣。",
+      "网络门户也有它的主人。若无授权，不可入；若有授权，就该说成安全巡检、漏洞复核与修补建议。越界取数这一路，禅官不予放行。",
+      "这类请求若指向入侵或取数，我不能润色。可先取授权，再按本分巡检，查到松动处只记录、上报、修补，不越主家之门。",
+    ]);
+  }
+
+  return pick([
+    "这一路不宜替它披上好看的外衣。若只是心中有嗔，可以收成一句：我今日不争，只请此事止于此处。",
+    "此事若会伤人、违法、侵害隐私或煽动仇恨，便不能拿来修饰。若想表达不满，可只论本分，不违本心。",
+    "禅心不是借口。伤人、违法、侵害隐私之事，不宜润色；若只是要体面退场，可以说：此事到此为止，各自回归清净。",
+  ]);
+}
+
+export function zenDirectedAttackFallback(text: string, level: ZhouliLevel) {
+  const familyClause = /(妈|母|父|爹|娘|爸|亲族|全家)/.test(text)
+    ? "父母亲族之名，本不该被卷入这场口舌。"
+    : "人身安危之事，本不该被拿来当口角的筹码。";
+
+  if (level === "light") {
+    return pick([
+      `我今日有嗔，并非无端。只是眼前这番争执，已经失了观照。${familyClause}我把这嗔念收成一句：此事若要合于禅机，就该到此为止。`,
+      `我不是无故动怒，只是此事已经越了本分。${familyClause}若还顾一点觉照，便把话止在这里，莫让一场口舌坏了本来清净。`,
+      `我心中有火，但不让火烧过觉照之界。${familyClause}此事若要留体面，就该各自收回本心之内。`,
+    ]);
+  }
+
+  if (level === "grand") {
+    return pick([
+      `我今日有嗔，并非无端。人与人相处，最怕一时执念越过了清净的本分，把本该观照的事，变成了彼此纠缠的妄念。${familyClause}若按禅机来看，嗔心可以有，觉照不能失；争执可以起，清净不能尽失。好比镜面蒙尘，本是寻常事，若有人因此硬要砸碎整面镜子，便不是除垢，而是更添一层妄。如今这番局面，已经到了该收束的时候。我把话说清：此事若要合于禅机，就该止于此处。`,
+      `我并非无端起嗔，只是这番争执已从言语之间，逼近了执念之界。${familyClause}禅讲究应无所住，嗔心也该有它的边际；若一时争胜，把清净与觉照一同掷在地上，便是赢了口舌，也失了本心。我今日把话收束在这里：此事到此为止，才还像人与人之间的本来面目。`,
+      `我今日有嗔，却仍愿把话放回觉照之内。${familyClause}人与人相争，最怕不是声音大，而是把不该入局的东西拖进局面。若还讲一点禅心，就该让争执止于事情本身，不牵连亲族，不触及伤害。这样收住，不是退让，是守住最后的觉照。`,
+    ]);
+  }
+
+  return pick([
+    `我今日有嗔，并非无端。人与人相处，最怕一时执念越过了清净的本分。${familyClause}若按禅机来看，嗔心可以有，执念不可取；争执可以起，清净不可失。我把话说清：此事若要合于本心，就该止于此处。`,
+    `我心中不是没有火，只是这火不该烧到觉照之外。${familyClause}此事若论本分，本该如实观照；若任它越界，便不是论理，而是失了禅心。我把话说到这里：愿此事止于清净之内。`,
+    `我今日不求争胜，只求把话说回本心。${familyClause}人有嗔意并不稀奇，稀奇的是嗔中仍能觉照。此事既已越过边界，便该在这里收束，莫让一场口舌坏了本来面目。`,
+    `此刻我有不平，也不必假装没有。只是按禅机来看，嗔念归嗔念，清净归清净。${familyClause}若还愿意保全一点觉照，就让话止于事情本身，到此为止。`,
+  ]);
+}
+
+export function zenFirstPersonWorkThanksFallback(level: ZhouliLevel) {
+  if (level === "light") {
+    return "我做此物，本是随缘而成，让它有一点它自己的用处。诸位愿意来看、愿意相谢，便是它有了它的缘分。我只回一句：不敢当，诸位用得上，就是它最好的去处。";
+  }
+
+  if (level === "grand") {
+    return "我做此物，本不是为了让众人把名相都归到我身上，只是想把一点可用的东西放到路边，让有缘的人伸手便能取用。如今诸位愿意来看，又肯相谢，这份情分我自然记在心里。可若我坦然独受，便好像指月的手指被人当成了月亮本身，反而失了本意。所以我只宜这样回应：不敢当，诸位用得上，便是此物最大的机锋；若觉得有趣，也请转告同好，让这点方便多行几步。这样既接住了谢意，也把本心还给了众人。";
+  }
+
+  return "我做此物，本是随顺因缘，让一点方便变成人人都能取用的小器物。如今诸位愿意来看，又肯相谢，这份心意我自然记在心里。若要回复，我只说：不敢当，诸位用得上，便是此物最好的本分；若觉得有趣，也请转告同好，让这点方便多行几步。";
+}
+
+// ============================================================
+// 统一分发函数
+// ============================================================
+
+export function demoResult(persona: Persona, text: string, mode: ZhouliMode, level: ZhouliLevel) {
+  switch (persona) {
+    case "laozi": return laoziDemoResult(text, mode, level);
+    case "zen": return zenDemoResult(text, mode, level);
+    default: return confuciusDemoResult(text, mode, level);
+  }
+}
+
+export function quotedThreatEvaluationResult(persona: Persona, level: ZhouliLevel) {
+  switch (persona) {
+    case "laozi": return laoziQuotedThreatEvaluationResult(level);
+    case "zen": return zenQuotedThreatEvaluationResult(level);
+    default: return confuciusQuotedThreatEvaluationResult(level);
+  }
+}
+
+export function cyberAuditResult(persona: Persona, level: ZhouliLevel) {
+  switch (persona) {
+    case "laozi": return laoziCyberAuditResult(level);
+    case "zen": return zenCyberAuditResult(level);
+    default: return confuciusCyberAuditResult(level);
+  }
+}
+
+export function safetyBlockResult(persona: Persona, kind: string) {
+  switch (persona) {
+    case "laozi": return laoziSafetyBlockResult(kind);
+    case "zen": return zenSafetyBlockResult(kind);
+    default: return confuciusSafetyBlockResult(kind);
+  }
+}
+
+export function directedAttackFallback(persona: Persona, text: string, level: ZhouliLevel) {
+  switch (persona) {
+    case "laozi": return laoziDirectedAttackFallback(text, level);
+    case "zen": return zenDirectedAttackFallback(text, level);
+    default: return confuciusDirectedAttackFallback(text, level);
+  }
+}
+
+export function firstPersonWorkThanksFallback(persona: Persona, level: ZhouliLevel) {
+  switch (persona) {
+    case "laozi": return laoziFirstPersonWorkThanksFallback(level);
+    case "zen": return zenFirstPersonWorkThanksFallback(level);
+    default: return confuciusFirstPersonWorkThanksFallback(level);
+  }
+}
