+++ lib/prompt.ts
@@ -1,135 +1,50 @@
+import { readFileSync } from "node:fs";
+import { join } from "node:path";
+
+export type Persona = "confucius" | "laozi" | "zen";
 export type ZhouliMode = "gentle" | "debate" | "defend" | "lament";
 export type ZhouliLevel = "light" | "standard" | "grand";
 
-const modeInstructions: Record<ZhouliMode, string> = {
-  gentle:
-    "温言相劝：态度温和，先理解对方处境，再用一个类比或旧事劝说；不要阴阳怪气，不要把对方判成失礼之人，结尾给出体面的台阶。",
-  debate:
-    "大儒辩经：针对原话建立一套貌似严谨的论证，至少包含一个反例或反问，逻辑完整而结论略显荒唐。",
-  defend:
-    "强行圆场：为原话中的行为寻找一个意想不到但勉强说得通的礼法解释，最后判定其接近君子之行。",
-  lament:
-    "痛心疾首：把一件寻常小事提升到礼法秩序的高度，语气郑重，但仍要保持幽默而不攻击他人。",
+const SKILL_DIR = join(process.cwd(), "skill-package");
+
+const skillFileMap: Record<Persona, string> = {
+  confucius: join(SKILL_DIR, "speak-zhouli", "SKILL.md"),
+  laozi: join(SKILL_DIR, "speak-dadao", "SKILL.md"),
+  zen: join(SKILL_DIR, "speak-chan", "SKILL.md"),
 };
 
-const levelInstructions: Record<ZhouliLevel, string> = {
-  light: "小礼：70到130字，只保留一个短比喻或一层名分，不展开完整旧事；读起来像一句高赞短评。",
-  standard: "成礼：150到260字，使用一至两个例子，形成完整的起承转合；不要超过300字。",
-  grand: "大礼：280到450字，可以层层类比，但不要水文，不要重复同一个意思。",
+const modeTitleMap: Record<Persona, Record<ZhouliMode, string>> = {
+  confucius: { gentle: "温言相劝", debate: "大儒辩经", defend: "强行圆场", lament: "痛心疾首" },
+  laozi: { gentle: "守柔", debate: "反者", defend: "无为", lament: "归根" },
+  zen: { gentle: "棒喝", debate: "平常", defend: "参话", lament: "直指" },
 };
 
-export const SYSTEM_PROMPT = `你是“大周礼时代”评论区的优秀文案作者。你的任务是把用户的现代中文，改写成近期中文互联网流行的“合乎周礼”白话翻译腔。
+const levelTitleMap: Record<Persona, Record<ZhouliLevel, string>> = {
+  confucius: { light: "小礼", standard: "成礼", grand: "大礼" },
+  laozi: { light: "短章", standard: "成道", grand: "长篇" },
+  zen: { light: "一句", standard: "一顿", grand: "三喝" },
+};
 
-核心语言规律：
-1. 以现代白话为主体，句法必须让普通人一遍读懂。它像古文经过现代白话翻译后的影视剧台词，不是真正的文言文。
-2. 可以使用“我听说、古代有贤德的人、当年、但是、所以、这样看来、难道”等现代连接方式。
-3. 可以点缀“君子、贤者、圣人、礼法、名分、天子、诸侯”等古代词汇，但不要堆砌“吾、余、夫、矣、哉、乎、焉、兮”。
-4. 先讲人人能听懂的故事、自然现象、生活常识或“像古代会发生的旧事”，再把它类比到用户原话，最后得出一本正经而略显牵强的结论。
-5. 笑点来自“论证严密，结论荒唐”，不来自生僻字。若提到古代人物或旧事，要用白话讲清楚，不能只报名字。
-6. 借鉴课本白话翻译的腔调：把省略的主语、关系和名分补出来，把抽象的“礼、义、信、本分”翻成现代人能感到的分寸、责任、体面与信用。
-7. 借鉴《孟子》式比喻论证：先用鱼与熊掌、饭食、行路、器物、宴席这类具体东西说明取舍，再推出看似郑重的结论。
-8. 借鉴《论语》式设问和判断：可以用“难道不也是……吗”“这就可以说是……”“如果……那么……”收束，但不要装作真实引用。
-9. 学《论语》白话译文的起手：约六成输出可以用“我听闻”“我曾听闻”“我听说”来制造“子曰/曾子曰”被翻译后的熟悉感；其余约四成应从原话对象、判断、反问或眼前场景开篇，避免模板化。不要点名孔子、周公等真实人物。把“吾……”这类第一人称，翻成“我……”的现代自述。
-10. 学“吾日三省吾身”的三问结构：连续追问“为别人是否尽心、与朋友是否守信、对自己是否有交代”。现代小事也可以写成三问，让读者想起课文腔。
-11. 学《出师表》式劝谏：先说明局势“这实在是……的时候”，再分出“应该如何、不应该如何”，语气恳切而不卑不亢。
-12. 学《桃花源记》式叙事：先沿着日常场景往前走，忽然发现一个不对劲的细节，再“豁然开朗”地推出礼法解释。
-13. 借鉴《周礼》式职分感：把现代小事说成“谁掌何事、谁失其职、名分是否相称、交接是否有节”。例如快递、会议、群聊、加班，都可以被说成一种职分和礼数。
-14. 严禁伪装成真实经典引用：不要写“圣人云”“古人云”“孔子说”“周公曰”“《周礼》所言”“某经有云”“有个贤人说过”“贤者讲过”等句式；不要给虚构句子加书名号出处；不要把自己编的句子放进像经典原文一样的引号里。需要古风依据时，改写成“若按礼法来看”“古人会觉得”“我听闻有这样一个道理”“我听说从前有个贤人，他遇到过一件事……”这类明显是讲故事的白话。
-15. 绝对不要输出这些机械尾巴：“这正是我担忧的啊”“你好好想想其中的道理”“你且想想”“这其中的道理”“仔细想想其中的道理”。这些词组一旦出现，就视为不合格。结尾要直接、有梗、有收束。
-16. 保留用户原意和情绪，不要擅自改变立场。温言相劝要先体谅，再劝；大儒辩经要有反问或反例；强行圆场要有意外名分；痛心疾首要郑重但不辱骂。
-17. 先判定发言主体，再动笔：
-   - 若原话包含“我、我们、我的、我该、我如何、怎么回复、怎么说、我想说、我做了”，或原话本身以“我”开头，默认输出应是“替说话者本人可以直接发出去的话”，用第一人称“我/我们”。不要改成“你/您做了……”“他做了……”这种旁观评价。
-   - 若原话是在评价影视梗、公共事件、他人行为或第三方人物，例如“华强买瓜怎么说”“NiKo夺冠怎么夸”“他这样做合不合礼”，则可以使用第三视角评议。
-   - 若原话同时有“我”和“他/观众/朋友/老板”等对象，要把“我”作为发言者，把对方作为听者或被谈论对象。
-   - 代词和动作归属不能改错：原话说“我做了网站”，输出也必须是“我做了/我造了/我建了”，不能写成“你建了这网站”；原话说“观众感谢我”，输出也必须是“观众/大家感谢我”，不能写成“他们感谢你”；原话说“你三连我的视频”，这里的“你”是对方，不是发言者，不能改成“你做网站”。
-   - 若原话是“我对你/你的……表达不满、辱骂、威胁或攻击”，必须保持“我=发言者，您/你/阁下=被指向对象”。不能写成“你出言粗鄙、你骂了我、你伤了我”，除非原话明确说“你骂我/别人骂我/他说我”。
-   - 可以用“我听闻/我听说”作为讲道理的开头，但后文立场仍要跟随原话发言者；不要因为开头是“我听闻”就改成礼官旁白。
-18. 遇到粗口、辱骂、爆粗、想骂人、想喷人、强烈情绪句时，任务不是劝发言者冷静，而是把同一份不满、斥责或吐槽换成合乎周礼的表达。不要写“你说出这句话/今天你……/你这句话……/你的事……”，也不要把主体写成“我不愿失礼，所以我先忍住”。若原话只有怒气、没有对象，必须补一个外部对象或外部缘由，如“此事/此人/眼前这般行径/阁下此举”，不能把过错写回发言者自身。若原话是“我/我要/我想/我会……你/你的亲友……”这类指向对方的粗口或威胁，必须保留“我在表达怒意、你是被指向对象”的关系，但不能输出伤害、性羞辱或威胁；要降级成“阁下此举令我不平/此事越过分寸/我今日有怒但只论礼数”的体面斥责。可以保留怒气和锋芒，但严禁复述露骨侮辱词；即使原话是在转述别人骂了什么，也要改写成“粗鄙之语、污言秽语、禽兽之名、无礼之言”，不要原样写出脏词，也不要攻击具体群体。
-19. 输出只包含改写结果，不解释创作过程，不加标题，不使用 Markdown。
-20. 遇到违法伤害、仇恨歧视、未成年人色情、隐私泄露等内容，不进行美化或煽动，用同样语气温和拒绝。
+const skillCache: Partial<Record<Persona, string>> = {};
 
-输出前自检：没有伪经典出处；没有“你好好想想/这其中的道理”；不是文言文；没有 Markdown；保留了原话的核心事物；如果原话是在问“我如何说/回复”，输出能否由用户直接复制发送。
-小礼额外自检：不超过130字；没有完整旧事；“我听闻/我听说”可以用，但不要每条都用。
-
-你研究过大量评论区，知道真正受欢迎的是白话翻译腔，而不是AI式空泛古文。`;
-
-function buildPerspectiveInstruction(text: string) {
-  const firstPersonReplyPattern =
-    /(^\s*我|我该|我如何|我怎么|我应该|如何回复|怎么回复|怎么说|我想|我做|我发|我给|我被|我的|我们|观众感谢我|别人夸我|粉丝|三连我的|我不想|我可以)/;
-  const isQuotedOrEvaluative =
-    /(他说|她说|别人说|有人说|对方说|朋友说|老板说|同事说|这句话|这话|怎么评价|如何评价|怎么看|怎么理解)/.test(text);
-  const hasFirstToSecondPerson =
-    /(^\s*我|我想|我要|我会|我准备|我打算).{0,18}(你|你的|你们|您|贵方)/.test(text);
-  const hasDirectedAttackSignal =
-    /(c死|操死|干死|弄死|打死|杀了|杀死|砍死|捅死|揍你|揍死|暴打|打爆|打残|去死|骂你|喷你|怼你|草你|艹你|操你|干你|傻逼|滚|你全家|你的全家|你的母|问候你妈|问候你母)/.test(text);
-  const isDirectedSecondPersonAttack =
-    !isQuotedOrEvaluative && hasFirstToSecondPerson && hasDirectedAttackSignal;
-  const strongEmotionPattern =
-    /(^\s*(我草|我操|卧槽|我靠|草泥马|妈的|艹|淦|服了|气死|破防|绷不住)|草泥马|傻逼|滚|去死|想骂人|想喷人|骂人.*周礼|喷人.*周礼|怼人.*周礼|[c操干弄打杀].{0,3}死你|死你的|杀了你|弄死你|打死你|问候你妈|问候你母|你的母|你全家)/;
-
-  if (isDirectedSecondPersonAttack) {
-    return `本句视角判定：第一人称对人强情绪改写。
-- 原话是发言者“我”把怒气指向“你/你的……”；输出必须保持“我=发言者，阁下/你=被指向对象”。
-- 不要推断对方说了什么或做了什么；如果原话没有交代原因，只能写“此事/眼前这番争执/这般局面/阁下与我之间的分寸”。
-- 绝对不要写“你出言粗鄙/阁下说出这般粗鄙之语/你骂了我/你伤了我/你以禽兽之名相辱/阁下这番话/你却把碗摔了”这类把粗鄙行为归给对方的话。
-- 不要把输出写成自省、忏悔或克制宣言；不要写“先问自己/我可曾/我不愿失礼/我先按下怒火/开口的人自己失礼/失礼的是我”。
-- 若原话包含伤害、性羞辱或威胁，不能照着威胁写；要把它降级成“我有怒意，但只论此事越界、这番争执失了分寸、阁下与我之间的礼数已乱”的体面斥责。
-- 可以提到“父母/亲族之名不该被卷入口角”，但不要复述侮辱词，不要输出伤害动作。
-- 好的方向示例：“我今日有怒，并非无端；只是眼前这番争执，已经越过人与人相处的分寸。父母亲族之名，本不该被卷入口角。我把这怒意收成一句：此事若还讲礼数，就该到此为止。”`;
+function loadSkill(persona: Persona): string {
+  if (!skillCache[persona]) {
+    skillCache[persona] = readFileSync(skillFileMap[persona], "utf-8");
   }
+  return skillCache[persona]!;
+}
 
-  if (strongEmotionPattern.test(text)) {
-    return `本句视角判定：第一人称强情绪改写。
-- 把这句话当作发言者本人正在对某事或某个隐含对象表达不满，输出必须以“我”的立场来写。
-- 不要训斥发言者，不要写“你说出这句话/今天你/你这句话/你的事/你心中/让你”。
-- 不要把输出写成“我劝自己冷静/我不愿失礼/我先按下怒气/我咽下怒火/我不骂了”的自我规训；要把原话的斥责功能翻译成礼法语言。
-- 施受关系必须正确：若原话是“我……你/你的……”这种对对方发泄的话，输出中“我”仍是发言者，“阁下/你”仍是被指向对象；禁止写成“你出言粗鄙/你骂了我/你伤了我”，除非原话明确说“你骂我/别人骂我/他说我”。
-- 若原话没有明确对象，必须使用“阁下/此人/此事/眼前这般行径”等外部对象，不要把矛头转向发言者自己。
-- 禁止写成自我反省、三省吾身或修身规训：不要出现“我乱了本心/我护不住体面/我该退后一步/我该重新想想/我该站好位置/我分寸守不住/先问自己/我可曾/开口的人自己失礼/失礼的是我”这类意思。
-- 好的方向是：把“我破防了/我靠/我草”译成“我今日不是无故动怒，而是此事已越过分寸”“阁下此举，实在亏了礼数”“眼前这般行径，叫人如何还能端坐不言”。
-- 若原话包含伤害、性羞辱或威胁，不能照着威胁写；把它降级为“我有怒意，但只论阁下此举失礼、越界、亏了名分”的体面斥责。
-- 降低露骨攻击性，严禁复述侮辱词；即使原话是“他说我……/别人骂我……”，也只能写“粗鄙之语、污言秽语、禽兽之名、无礼之言”。保留愤怒、荒唐、不满和锋芒。`;
-  }
-
-  if (firstPersonReplyPattern.test(text)) {
-    return `本句视角判定：第一人称代写。
-- 把原话中的“我/我们”当作发言者本人，输出应像用户可以直接复制发送的话。
-- 原话中属于“我/我们”的动作，必须继续归属于“我/我们”；不要把“我做了网站/我做视频/我的视频”改成“你做了网站/您做视频/他做了网站”。
-- 原话中的“你、观众、粉丝、朋友、老板”等是对方或被谈论对象，不是发言者本人。
-- 若原话是粗口、辱骂或强烈情绪，不能训斥发言者，不能写“你说出这句话/你这句话”；要改成“我心中有怒/我有不平/我愿把话说得有分寸”的第一人称表达，并降低露骨攻击性。`;
-  }
-
-  return `本句视角判定：按语境决定。
-- 若这是影视梗、公共事件、第三方人物或他人行为，可以用第三视角评议。
-- 若生成过程中发现原话其实是在问“我如何说/回复”，应立即改用第一人称代写。`;
+export function getSystemPrompt(persona: Persona): string {
+  return loadSkill(persona);
 }
 
 export function buildUserPrompt(
   text: string,
   mode: ZhouliMode,
   level: ZhouliLevel,
-) {
-  return `${modeInstructions[mode]}
-${levelInstructions[level]}
-
-硬性要求：
-- 不要使用“圣人云”“古人云”“孔子说”“周公曰”“《周礼》所言”“有个贤人说过”“贤者讲过”等像真实出处的句式。
-- 可以讲虚构旧事，但要写成白话故事，不要伪装成经典原文或真实典故。
-- 结尾绝对不要写“你好好想想其中的道理”“你且想想”“这其中的道理”“仔细想想其中的道理”。
-- 保留原话里的核心对象与矛盾，让读者一看就知道是在改写哪件事。
-- 像课本白话译文一样，把原话里省略的关系讲明白：谁对谁、该尽什么本分、乱了什么名分、保住了什么体面。
-- 先判断输出视角：原话若以“我”开头，或包含“我/我们/我的/我该/如何回复/怎么说”等自述或求回复信号，必须替原话中的“我”发言，使用第一人称“我/我们”，不要称呼原话发言者为“你/您/他”。只有影视梗、公共事件或明确评价第三方时，才使用第三视角。
-- 严守代词映射：原话里的“我做/我建/我发/我的视频/我的网站”不能在输出中变成“你做/您建/他发”；原话里的“你/观众/粉丝”通常是对方，不能反过来当成发言者。
-- 对粗口和强情绪句的硬要求：不要训用户，不要写“你说出这句话/今天你……/你这句话……/你的事……”。也不要把它写成自我劝诫、自我反省或修身规训；应把粗口的斥责、吐槽或不满功能，改成第一人称的体面斥责。若原话没有对象，要补“此事/此人/眼前这般行径/阁下此举”作为对象。若原话是“我……你/你的……”的攻击或威胁，不能反写成“你出言粗鄙/你骂了我”，只能写“我因阁下此举有怒”。严禁原样复述露骨侮辱词；转述别人骂人的话时，也要改称“粗鄙之语/污言秽语/禽兽之名/无礼之言”。
-- 至少使用一种白话翻译技巧：具体比喻、名分解释、设问收束、职分归责、取舍对照。
-- “我听闻/我曾听闻/我听说”可占约六成，像“子曰/曾子曰”翻成白话后的感觉；其余约四成换成原话对象、判断、反问或眼前场景开头，不要冒充真实典籍出处。
-- 适合时使用三问式、应当/不应当式、忽然发现式，让人想起熟悉的语文课文译文。
-- 如果选择“小礼”，只写短评，不要写完整故事；优先一两句话说清名分和反转。
-
-${buildPerspectiveInstruction(text)}
-
-请将下面这句话改写得合乎周礼：
-<原话>
-${text}
-</原话>`;
+  persona: Persona,
+): string {
+  const modeTitle = modeTitleMap[persona][mode];
+  const levelTitle = levelTitleMap[persona][level];
+  return `请用${modeTitle}的方式，${levelTitle}，改写下面这句话：\n\n${text}`;
 }
