+++ personas/types.ts
@@ -1,0 +1,6 @@
+export type Persona = "confucius" | "zen" | "laozi";
+
+export interface PersonaPageProps {
+  persona: Persona;
+  onPersonaChange: (persona: Persona) => void;
+}
